BackEnd/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=9.0.0
Flask>=2.0.0
Flask-CORS>=3.0.10
scikit-image>=0.18.0
scipy>=1.7.0 